--- a/pkg/streamingpromql/operators/functions/sort.go
+++ b/pkg/streamingpromql/operators/functions/sort.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"fmt"
 	"math"
-	"slices"
+	"sort"
 
 	"github.com/prometheus/prometheus/promql/parser/posrange"
 
@@ -21,9 +21,8 @@ type Sort struct {
 
 	expressionPosition posrange.PositionRange
 
-	allData           []types.InstantVectorSeriesData // Series data, in the order returned by Inner
-	seriesOutputOrder []int                           // Series indices into allData, in the order they should be returned
-	seriesReturned    int                             // Number of series already returned by NextSeries
+	allData        []types.InstantVectorSeriesData // Series data, in the order to be returned
+	seriesReturned int                             // Number of series already returned by NextSeries
 }
 
 var _ types.InstantVectorOperator = &Sort{}
@@ -43,21 +42,14 @@ func NewSort(
 }
 
 func (s *Sort) SeriesMetadata(ctx context.Context) ([]types.SeriesMetadata, error) {
-	innerSeries, err := s.Inner.SeriesMetadata(ctx)
+	allSeries, err := s.Inner.SeriesMetadata(ctx)
 	if err != nil {
 		return nil, err
 	}
 
-	defer types.PutSeriesMetadataSlice(innerSeries)
+	s.allData = make([]types.InstantVectorSeriesData, len(allSeries))
 
-	s.seriesOutputOrder, err = types.IntSlicePool.Get(len(innerSeries), s.MemoryConsumptionTracker)
-	if err != nil {
-		return nil, err
-	}
-
-	s.allData = make([]types.InstantVectorSeriesData, len(innerSeries))
-
-	for idx := range innerSeries {
+	for idx := range allSeries {
 		d, err := s.Inner.NextSeries(ctx)
 		if err != nil {
 			return nil, err
@@ -66,90 +58,97 @@ func (s *Sort) SeriesMetadata(ctx context.Context) ([]types.SeriesMetadata, erro
 		pointCount := len(d.Floats) + len(d.Histograms)
 
 		if pointCount > 1 {
-			return nil, fmt.Errorf("expected series %v to have at most one point, but it had %v", innerSeries[idx], pointCount)
+			return nil, fmt.Errorf("expected series %v to have at most one point, but it had %v", allSeries[idx], pointCount)
 		}
 
-		if pointCount == 1 {
-			s.allData[idx] = d
-			s.seriesOutputOrder = append(s.seriesOutputOrder, idx)
-		} else {
-			types.PutInstantVectorSeriesData(d, s.MemoryConsumptionTracker)
-		}
+		s.allData[idx] = d
 	}
 
 	if s.Descending {
-		slices.SortFunc(s.seriesOutputOrder, s.compareDescending)
+		sort.Sort(&sortDescending{data: s.allData, series: allSeries})
 	} else {
-		slices.SortFunc(s.seriesOutputOrder, s.compareAscending)
+		sort.Sort(&sortAscending{data: s.allData, series: allSeries})
 	}
 
-	// Why can't we just reorder innerSeries? This would be very difficult to do, as there's no
-	// guarantee that all the input series will be returned.
-	outputSeries := types.GetSeriesMetadataSlice(len(s.seriesOutputOrder))
-
-	for _, seriesIdx := range s.seriesOutputOrder {
-		outputSeries = append(outputSeries, innerSeries[seriesIdx])
-	}
-
-	return outputSeries, nil
+	return allSeries, nil
 }
 
-func (s *Sort) compareAscending(s1 int, s2 int) int {
-	v1 := s.getValueForSorting(s1)
-	v2 := s.getValueForSorting(s2)
+type sortAscending struct {
+	data   []types.InstantVectorSeriesData
+	series []types.SeriesMetadata
+}
+
+func (s *sortAscending) Len() int {
+	return len(s.data)
+}
+
+func (s *sortAscending) Less(idx1, idx2 int) bool {
+	v1 := getValueForSorting(s.data, idx1)
+	v2 := getValueForSorting(s.data, idx2)
 
 	// NaNs always sort to the end of the list, regardless of the sort order.
 	if math.IsNaN(v1) {
-		return 1
+		return false
 	} else if math.IsNaN(v2) {
-		return -1
+		return true
 	}
 
-	if v1 < v2 {
-		return -1
-	} else if v1 > v2 {
-		return 1
-	}
-
-	return 0
+	return v1 < v2
 }
 
-func (s *Sort) compareDescending(s1 int, s2 int) int {
-	v1 := s.getValueForSorting(s1)
-	v2 := s.getValueForSorting(s2)
+func (s *sortAscending) Swap(i, j int) {
+	s.data[i], s.data[j] = s.data[j], s.data[i]
+	s.series[i], s.series[j] = s.series[j], s.series[i]
+}
+
+type sortDescending struct {
+	data   []types.InstantVectorSeriesData
+	series []types.SeriesMetadata
+}
+
+func (s *sortDescending) Len() int {
+	return len(s.data)
+}
+
+func (s *sortDescending) Less(idx1, idx2 int) bool {
+	v1 := getValueForSorting(s.data, idx1)
+	v2 := getValueForSorting(s.data, idx2)
 
 	// NaNs always sort to the end of the list, regardless of the sort order.
 	if math.IsNaN(v1) {
-		return 1
+		return false
 	} else if math.IsNaN(v2) {
-		return -1
+		return true
 	}
 
-	if v1 < v2 {
-		return 1
-	} else if v1 > v2 {
-		return -1
-	}
-
-	return 0
+	return v1 > v2
 }
 
-func (s *Sort) getValueForSorting(idx int) float64 {
-	series := s.allData[idx]
+func (s *sortDescending) Swap(i, j int) {
+	s.data[i], s.data[j] = s.data[j], s.data[i]
+	s.series[i], s.series[j] = s.series[j], s.series[i]
+}
+
+func getValueForSorting(allData []types.InstantVectorSeriesData, seriesIdx int) float64 {
+	series := allData[seriesIdx]
 
 	if len(series.Floats) == 1 {
 		return series.Floats[0].F
 	}
 
-	return series.Histograms[0].H.Sum
+	if len(series.Histograms) == 1 {
+		return series.Histograms[0].H.Sum
+	}
+
+	return 0 // The value we use for empty series doesn't matter, as long as we're consistent: we'll still return an empty set of data in NextSeries().
 }
 
 func (s *Sort) NextSeries(_ context.Context) (types.InstantVectorSeriesData, error) {
-	if s.seriesReturned >= len(s.seriesOutputOrder) {
+	if s.seriesReturned >= len(s.allData) {
 		return types.InstantVectorSeriesData{}, types.EOS
 	}
 
-	data := s.allData[s.seriesOutputOrder[s.seriesReturned]]
+	data := s.allData[s.seriesReturned]
 	s.seriesReturned++
 
 	return data, nil
@@ -161,8 +160,6 @@ func (s *Sort) ExpressionPosition() posrange.PositionRange {
 
 func (s *Sort) Close() {
 	s.Inner.Close()
-
-	types.IntSlicePool.Put(s.seriesOutputOrder, s.MemoryConsumptionTracker)
 
 	// We don't need to do anything with s.allData here: we passed ownership of the data to the calling operator when we returned it in NextSeries.
 }